requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
socksio>=1.0.0  # SOCKS5 proxy support for httpx (LLM_PROXY_URL)
websockets>=12.0
pydantic-settings>=2.1.0
SQLAlchemy>=2.0.43
asyncpg>=0.31.0
psycopg2-binary>=2.9.9
alembic>=1.16.5
redis>=5.0.0
python-dotenv>=1.1.1
fakeredis>=2.33.0
pyyaml>=6.0.2
greenlet>=3.2.4
pgvector>=0.4.2
gunicorn>=23.0.0
scalar-fastapi>=1.6.1
mkdocs-material>=9.7.1
mkdocstrings[python]>=0.24.0
mkdocs-mermaid2-plugin>=1.2.3
mkdocs-static-i18n>=1.3.0
Scrapy>=2.13.4
pika>=1.3.2
prometheus-fastapi-instrumentator>=7.0.0
aiosqlite>=0.20.0
posthog>=3.5.0
aiogram>=3.17.0
aio-pika>=9.5.5
email-validator>=2.0.0
anthropic
google-genai
groq
strawberry-graphql[fastapi]

# Tracing (OpenTelemetry -> Tempo)
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-requests>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0

nats-py>=2.7.2
